% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheshire-petitioners.R
\docType{data}
\name{cheshire_petitioners}
\alias{cheshire_petitioners}
\title{TPOP Cheshire Petitioners}
\format{
A data frame with 1996 observations of 8 variables.
}
\usage{
cheshire_petitioners
}
\description{
Data about individuals named in petitions to Cheshire magistrates 1573-1798.
Includes name, name type, signature type, gender, residence, status, and petition_id to link to cheshire_petitions table.
}
\references{
The Power of Petitioning Data. https://doi.org/10.5281/zenodo.7027692
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
