% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheshire-petitions.R
\docType{data}
\name{cheshire_petitions}
\alias{cheshire_petitions}
\title{TPOP Cheshire Petitions}
\format{
A data frame with 613 observations of 9 variables.
}
\usage{
cheshire_petitions
}
\description{
Data about petitions to Cheshire magistrates 1573-1798.
Includes year, topic, magistrates' response, gender and petition type.
}
\references{
The Power of Petitioning Data. https://doi.org/10.5281/zenodo.7027692
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
