% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coroners.R
\docType{data}
\name{coroners}
\alias{coroners}
\title{Westminster Coroners' Inquests 1760-1799}
\format{
A data frame with 2885 observations of 18 variables.
}
\usage{
coroners
}
\description{
Data about coroners' inquests derived from the London Lives project.
Includes date, parish, deceased gender, cause of death, verdict.
}
\references{
London Lives Coroners' Inquests. https://doi.org/10.5281/zenodo.4402837
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
