% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obv-trials.R
\docType{data}
\name{obv_trials}
\alias{obv_trials}
\title{OBV trials 1780-1879}
\format{
A data frame with 20,593 observations of 28 variables.
}
\usage{
obv_trials
}
\description{
Summary data about Old Bailey Online trials with Old Bailey Corpus speech data (single-defendant trials only).
Includes date, counts of utterances and word counts for trials and for defendants; defendant name, gender, age, occupation; information about offence, verdict and sentence. Links to Old Bailey Online through defendant/trial IDs.
}
\references{
Old Bailey Voices text data (2018). https://doi.org/10.5281/zenodo.1143039
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
