% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-rhc.R
\docType{data}
\name{rhc}
\alias{rhc}
\title{DP Register of Habitual Criminals 1881-1925}
\format{
A data frame with 135996 observations of 11 variables.
}
\usage{
rhc
}
\description{
Selected data about individuals listed in the Registers of Habitual Criminals which were digitised by the Digital Panopticon project.
Includes year (of release), year of birth, gender, height in inches, marital status, physical build. IDs link to Digital Panopticon (www.digitalpanopticon.org/). Nominal record linkage was carried out on the register; many individuals appear multiple times and can be traced using the dp_life_id variable.
}
\references{
Metropolitan Police Register of Habitual Criminals, 1881-1925 (2018). https://doi.org/10.15131/shef.data.5697994.v1
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
