% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sab-dp-injuries.R
\docType{data}
\name{sab_dp_injuries}
\alias{sab_dp_injuries}
\title{Skin and Bone: Digital Panopticon injuries data 1790-1901}
\format{
A data frame with 68525 observations of 16 variables.
}
\usage{
sab_dp_injuries
}
\description{
Selected data about the injuries of individuals described in criminal records sourced from the Digital Panopticon project.
One injury per row: includes year of description, year of birth, gender, injury category and location. Person IDs link to Digital Panopticon (www.digitalpanopticon.org/). The description_id variable links together injuries in the same description. Some individuals appear multiple times and can be traced using the person_id variable.
}
\references{
Skin and Bone (2023). https://www.dhi.ac.uk/data/skin-and-bone
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
