% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sab-hp-injuries.R
\docType{data}
\name{sab_hp_injuries}
\alias{sab_hp_injuries}
\title{Skin and Bone: Hospital records injuries data 1790-1901}
\format{
A data frame with 17848 observations of 18 variables.
}
\usage{
sab_hp_injuries
}
\description{
Selected data about the injuries of individuals described in hospital records sourced from Madeleine Mant.
One injury per row: includes year of description, year of birth, gender, injury category and location.  The description_id variable links together injuries in the same description.
}
\references{
Skin and Bone (2023). https://www.dhi.ac.uk/data/skin-and-bone
}
\author{
\href{https://sharonhoward.org}{Sharon Howard}
}
\keyword{datasets}
